+++ hoot-core-test/src/test/cpp/hoot/core/algorithms/extractors/poi-polygon/PoiPolygonAlphaShapeDistanceExtractorTest.cpp
