+++ hoot-core-test/src/test/cpp/hoot/core/criterion/poi-polygon/PoiPolygonPoiCriterionTest.cpp
@@ -27,7 +27,7 @@
 
 // Hoot
 #include <hoot/core/TestUtils.h>
-#include <hoot/core/conflate/poi-polygon/criterion/PoiPolygonPoiCriterion.h>
+#include <hoot/core/criterion/poi-polygon/PoiPolygonPoiCriterion.h>
 
 using namespace geos::geom;
