+++ hoot-core-test/src/test/cpp/hoot/core/criterion/poi-polygon/PoiPolygonPolyCriterionTest.cpp
@@ -27,7 +27,7 @@
 
 // Hoot
 #include <hoot/core/TestUtils.h>
-#include <hoot/core/conflate/poi-polygon/criterion/PoiPolygonPolyCriterion.h>
+#include <hoot/core/criterion/poi-polygon/PoiPolygonPolyCriterion.h>
 
 using namespace geos::geom;
