+++ hoot-core-test/src/test/cpp/hoot/core/elements/NodeReplacementsTest.cpp
