+++ hoot-core-test/src/test/cpp/hoot/core/visitors/poi-polygon/PoiPolygonMatchVisitorTest.cpp
