+++ hoot-core/src/main/cpp/hoot/core/algorithms/extractors/poi-polygon/PoiPolygonAlphaShapeDistanceExtractor.cpp
