+++ hoot-core/src/main/cpp/hoot/core/conflate/poi-polygon/PoiPolygonRfClassifier.cpp
@@ -27,11 +27,11 @@
 #include "PoiPolygonRfClassifier.h"
 
 // hoot
-#include <hoot/core/conflate/poi-polygon/extractors/PoiPolygonAddressScoreExtractor.h>
-#include <hoot/core/conflate/poi-polygon/extractors/PoiPolygonAlphaShapeDistanceExtractor.h>
-#include <hoot/core/conflate/poi-polygon/extractors/PoiPolygonDistanceExtractor.h>
-#include <hoot/core/conflate/poi-polygon/extractors/PoiPolygonNameScoreExtractor.h>
-#include <hoot/core/conflate/poi-polygon/extractors/PoiPolygonTypeScoreExtractor.h>
+#include <hoot/core/algorithms/extractors/poi-polygon/PoiPolygonAddressScoreExtractor.h>
+#include <hoot/core/algorithms/extractors/poi-polygon/PoiPolygonAlphaShapeDistanceExtractor.h>
+#include <hoot/core/algorithms/extractors/poi-polygon/PoiPolygonDistanceExtractor.h>
+#include <hoot/core/algorithms/extractors/poi-polygon/PoiPolygonNameScoreExtractor.h>
+#include <hoot/core/algorithms/extractors/poi-polygon/PoiPolygonTypeScoreExtractor.h>
 #include <hoot/core/util/Factory.h>
 
 using namespace std;
