+++ hoot-core/src/main/cpp/hoot/core/elements/NodeReplacements.cpp
