+++ hoot-core/src/main/cpp/hoot/core/elements/NodeToWayMap.cpp
