+++ hoot-core/src/main/cpp/hoot/core/ops/HighwayReviewCleanerOp.h
