+++ hoot-core/src/main/cpp/hoot/core/ops/RemoveRoundabouts.h
