+++ hoot-core/src/main/cpp/hoot/core/ops/SmallWayMerger.h
