+++ test-files/ops/SmallWayMerger/SmallWayMergerInput1.osm
