+++ test-files/ops/SmallWayMerger/SmallWayMergerInput2.osm
